BOL-375 refactor: member 관련 model(in data module) package 분리
