+++ data/src/main/java/com/yapp/bol/data/datasource/member/MemberDataSource.kt
@@ -1,8 +1,8 @@
 package com.yapp.bol.data.datasource.member
 
 import com.yapp.bol.data.model.base.ErrorResponse
-import com.yapp.bol.data.model.group.MemberListResponse
-import com.yapp.bol.data.model.group.response.MemberValidApiResponse
+import com.yapp.bol.data.model.member.MemberListResponse
+import com.yapp.bol.data.model.member.MemberValidApiResponse
 import com.yapp.bol.domain.model.ApiResult
 import kotlinx.coroutines.flow.Flow
