+++ data/src/main/java/com/yapp/bol/data/model/member/GuestAddApiRequest.kt
@@ -1,4 +1,4 @@
-package com.yapp.bol.data.model.group
+package com.yapp.bol.data.model.member
 
 data class GuestAddApiRequest(
     val nickname: String
